+++ .ls-lint.yml
@@ -13,7 +13,6 @@ ignore:
   - 'crates/rolldown/tests/fixtures/**/_test.js' # `_test.js` is a special file, which will be treat as the entry file while executing the build output of fixture.
   - packages/rolldown/build.config.ts # convention name for using unbuild
   - packages/rolldown/vitest.config.ts
-  - packages/cli/test/**/*.test.js
   - web/docs/.vitepress/cache # cache files generated by vitepress
   # FIXME: should not ignore following folders
   - web
