--- packages/cli/bin/cli.js
@@ -1,2 +1,0 @@
-#!/usr/bin/env node
-import('../lib/main.js')
