--- packages/cli/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "rolldown-cli",
-  "version": "0.0.5",
-  "description": "Command-line interface for rolldown",
-  "keywords": [],
-  "author": "",
-  "license": "MIT",
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org/",
-    "access": "public"
-  },
-  "main": "index.js",
-  "type": "module",
-  "files": [
-    "lib",
-    "bin"
-  ],
-  "bin": {
-    "rolldown": "./bin/cli.js"
-  },
-  "scripts": {
-    "test": "node --test test/*.test.js"
-  },
-  "dependencies": {
-    "citty": "^0.1.6",
-    "consola": "^3.2.3",
-    "jiti": "^1.21.0",
-    "rolldown": "workspace:*"
-  }
-}
