--- packages/cli/test/fixtures/package.json
@@ -1,3 +1,0 @@
-{
-  "name": "invalid-cli-metadata"
-}
