+++ packages/rolldown/bin/cli.js
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+import('../cli/main.js')
