+++ packages/rolldown/cli/errors.js
@@ -1,3 +1,2 @@
-export const ERR_CLI_META_DATA = 'cli meta data error'
 export const ERR_UNSUPPORTED_CONFIG_FORMAT =
   "Unsupported config format. please use '.js', '.mjs' and '.ts' format"
