+++ packages/rolldown/package.json
@@ -13,9 +13,14 @@
     "rolldown"
   ],
   "files": [
+    "bin",
     "dist",
     "!dist/*.node"
   ],
+  "bin": {
+    "rolldown": "./bin/cli.js"
+  },
+  "type": "module",
   "main": "./dist/index.cjs",
   "types": "./dist/index.d.ts",
   "exports": {
@@ -60,6 +65,11 @@
       "aarch64-pc-windows-msvc",
       "armv7-linux-androideabi"
     ]
+  },
+  "dependencies": {
+    "citty": "^0.1.6",
+    "consola": "^3.2.3",
+    "jiti": "^1.21.0"
   },
   "devDependencies": {
     "@napi-rs/cli": "^3.0.0-alpha.43",
