+++ packages/rolldown/test/fixtures/rolldown.config.ts
