+++ packages/rolldown/vitest.config.mts
@@ -2,7 +2,7 @@ import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
   test: {
-    include: ['./test/runner.ts'],
+    include: ['./test/runner.ts', './test/*.test.js'],
     testTimeout: 20000,
   },
   esbuild: {
